refactor: rename folder albums
