+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UserModule } from './modules/users/users.module';
-import { AlbumModule } from './modules/album/albums.module';
+import { AlbumModule } from './modules/albums/albums.module';
 
 @Module({
   imports: [UserModule, AlbumModule],
