+++ src/db/in-memory.db.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 import { UserModel } from '../modules/users/user.model';
-import { AlbumModel } from '../modules/album/album.model';
+import { AlbumModel } from '../modules/albums/album.model';
 
 @Injectable()
 export class InMemoryDb {
