+++ src/modules/albums/albums.module.ts
